refactor(surface): Move code to do propagation to cells into its own module so it can be reused for
