+++ lib/get-surface.js
@@ -1,8 +1,6 @@
-import fill from 'lodash.fill'
 import dbg from 'debug'
+import propagate from './propagation'
 const debug = dbg('browsochrones:get-surface')
-
-import {getNonTransitTime, ITERATION_WIDTH} from './origin'
 
 /**
  * Get a travel time surface and accessibility results for a particular origin.
@@ -21,63 +19,28 @@ export default function getSurface ({origin, query, stopTreeCache, which}) {
 
   // how many departure minutes are there. skip number of stops
   const nMinutes = origin.data[transitOffset + 1]
-  const travelTimesForDest = new Uint8Array(nMinutes) // the total travel time per iteration to reach a particular destination
-  const waitTimesForDest = new Uint8Array(nMinutes) // wait time per iteration for particular destination
-  const inVehicleTravelTimesForDest = new Uint8Array(nMinutes) // in-vehicle travel time per destination
-  const walkTimesForDest = new Uint8Array(nMinutes)
 
-  // x and y refer to pixel not origins here
-  // loop over rows first
-  for (let y = 0, pixelIdx = 0, stcOffset = 0; y < query.height; y++) {
-    for (let x = 0; x < query.width; x++, pixelIdx++) {
-      const nStops = stopTreeCache.data[stcOffset++]
-
-      // can we reach this pixel without riding transit?
-      const nonTransitTime = getNonTransitTime(origin, {x, y})
-
-      // fill with unreachable, or the walk distance
-      fill(travelTimesForDest, nonTransitTime)
-      fill(waitTimesForDest, 255)
-      fill(inVehicleTravelTimesForDest, 255)
-      fill(walkTimesForDest, nonTransitTime) // when the origin is within walking distance and
-        // walking is the fastest way to reach the destination, _everything_ is walk time
-
-      for (let stopIdx = 0; stopIdx < nStops; stopIdx++) {
-        // read the stop ID
-        const stopId = stopTreeCache.data[stcOffset++]
-
-        // read the time (minutes)
-        const time = stopTreeCache.data[stcOffset++]
-
-        for (let minute = 0; minute < nMinutes; minute++) {
-          const offset = origin.index[stopId] + minute * ITERATION_WIDTH
-          const travelTimeToStop = origin.data[offset]
-
-          if (travelTimeToStop !== -1) {
-            const travelTimeToPixel = travelTimeToStop + time
-
-            // no need to check that travelTimeToPixel < 255 as travelTimesForDest[minute] is preinitialized to the nontransit time or 255
-            if (travelTimesForDest[minute] > travelTimeToPixel) {
-              travelTimesForDest[minute] = travelTimeToPixel
-              let inVehicle = inVehicleTravelTimesForDest[minute] = origin.data[offset + 1]
-              let wait = waitTimesForDest[minute] = origin.data[offset + 2]
-              // NB when we're talking about a particular trip, then walk + wait + inVehicle == total
-              // However if you calculate summary statistics for each of these individually, that may
-              // not be true. So we need to calculate walk here and explicitly calculate summary stats about it.
-              let walkTime = travelTimeToPixel - wait - inVehicle
-              walkTimesForDest[minute] = walkTime
-            }
-          }
-        }
-      }
-
+  propagate({
+    query,
+    stopTreeCache,
+    origin,
+    nMinutes,
+    callback: ({
+      travelTimesForDest,
+      walkTimesForDest,
+      inVehicleTravelTimesForDest,
+      waitTimesForDest,
+      x,
+      y
+    }) => {
+      const pixelIdx = y * query.width + x
       // compute and set value for pixel
       surface[pixelIdx] = computePixelValue(which, travelTimesForDest)
       waitTimes[pixelIdx] = computePixelValue(which, waitTimesForDest)
       walkTimes[pixelIdx] = computePixelValue(which, walkTimesForDest)
       inVehicleTravelTimes[pixelIdx] = computePixelValue(which, inVehicleTravelTimesForDest)
     }
-  }
+  })
 
   debug('generating surface complete')
